_CPUFreq: fix min/max/base/turbo frequencies on HWP systems
On HWP systems, the correct way of getting Pm, Pn, P1, and P01 frequencies is
via 'MSR_HWP_CAPABILITIES', not via 'MSR_PLATFORM_INFO'. In many cases they are
the same, but not always. E.g., Intel SST feature may change CPU base
frequency, but this will not be reflected in 'MSR_PLATFORM_INFO', while it will
be reflected in 'MSR_HWP_CAPABILITIES'.

Address the issue.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>