Tpmi: fix register bitfield write functionality
This implementation had a couple of bugs in it. First, the current value of
the register should be read in full, not just the named bitfield. Otherwise
the bitwise operation following the read will wipe the register content
except for the new field. Secondly, the bitwise operation should also clear
the field content first, otherwise it will only be able to enable bits.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>