_UncoreFreq: fix the legacy sysfs workaround
When unlocking the new sysfs API, not only remove all the legacy interface
limits, but also remember and restore the current values.

This fixes the following issue on GNR platform.

$  pepc pstates config --max-uncore-freq 1GHz
Max. uncore frequency: set to 1GHz for all dies in all packages
$.pepc pstates config --min-uncore-freq 1GHz
Min. uncore frequency: set to 1GHz for all dies in all packages
$ pepc pstates info --min-uncore-freq --max-uncore-freq
Min. uncore frequency: 1GHz for all dies in all packages
Max. uncore frequency: 2.5GHz for all dies in all packages

Notice that max. uncore frequency is 2.5GHz at the end, instead of expected
1GHz.

The bug is related to the uncore driver legacy API workaroun. When unlocking
it, the frequency should be restored to the previous value, but it was not. Fix
this.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>