_CPUFreq: add support for 'verify' config option
By default, the verification for written values is enabled. To disable,
set the 'verify' option to False. In this case, no read-back of the
values is done from the sysfs. Instead, some other mechanism should be
used to verify the result, like reading the current frequency back from
the sysfs.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>