pepc: fix '--cores all' and '--dies all' failures
Address the following failure:

$ pepc pstates config --min-uncore-freq min --dies all
pepc: error: BUG: no package and die numbers provided for setting Min. uncore frequency

(and a similar failure in case of '--cores all')

The problem is that because core and die numbers may be relative to package
numbers, and packages were not specified, the '_OpTarget.py' module tries to be
smart and resolve possible ambiguity in '_build_package_indexed_dict()'.
However, this method does not handle the "all" value correctly.

Fix this problem by adding an exception and just assuming '--packages all' in
the above cases.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>