_OpTarget: fix core and die numbers validation
Fix validation of core and die numbers and verify that all the specified
core/die numbers actually exist in any of the packages.

This fix only affects the case when core and die numbers are provided without
specifying package numbers, and '_build_package_indexed_dict()' is used to
resolve the ambiguity. This method is called before 'nums' are validated, so it
may include invalid numbers. Enhance the method to validate the input numbers
in 'nums'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>