tests: remove '--override-cpu-model' option coverage
The '--override-cpu-model' option is good to have, and nice to test, but it
cannot be uses on some systems, because it confuses CPU dopology discovery. For
example, when overriding SRF system with SPR CPUs, die numbers get messed up.

Remove it from tests. At some point we may remove it completely, it has not
been useful for really long time.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>