helperlibs: ProcessManager: fix occasional SSH process failures
Symptom:
pepclibs.helperlibs.Exceptions.Error: Expected only one line with PID in stdout and no lines in stderr, got 1 lines in stderr instead.

The problem is that the stderr output line may be fetched before the stdout
line, and returned to 'read_pid()' along with the PID in stdout.

Add -1 to 'lines' to prevent situations like this.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>