SSHProcessManager: fix occasional failures by using 2 markers
SSHProcessManager sometimes loses the last line(s) of stderr, which sometimes
may lead to errors (e.g., when the command was not found, and the "command not
found" message from stderr gets lost).

Fix this by having the marker on both stdout and stderr. And wait for both
before deciding that the program has finished.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>