_ProcessManagerBaser: remove must_exist argument
The 'must_exist' argument in read(), lsdir() and abspath() is changing the
return type of the methods so that they may return None whin must_exist is
False. This makes type checking and annotations a lot harder.

Notice that must_exist=False is not used much, and remove it. Also rename
read() to read_file().

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>