Pstates: change --frequencies mechanism from 'doc' to 'sysfs'
On Intel platforms with the intel_pstate driver, the --frequencies option
previously used the 'doc' mechanism to indicate that all frequencies between
min and max (with a 100MHz step) are supported. However, the minimum and
maximum frequencies are actually retrieved from sysfs, and only the 100MHz
step is based on documentation.

Therefore, the correct mechanism is 'sysfs'. Update the mechanism name
from 'doc' to 'sysfs' for clarity.

Fixes: https://github.com/intel/pepc/issues/10

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>