pepclibs: msr: Deep copy the FEATURES dict
At least in testing, sometimes it happens that the module is once imported, and
then used for various platforms. If the module FEATURES dictionary is modified
by the first platform, it does not get adjusted for the second platform. This
causes subtle bugs.

Prevent this by never touching FEATURES dict.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>